Remove get_isin
Removing this, because (a) it's not a Yahoo API it's hitting, and (b) it
doesn't return values for things I want to query.

It wouldn't necessarily be a problem for it to not be a Yahoo API, but I want
to sort out rate limiting before introducing other endpoints, and this one will
complicate things.